requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-anthropic==0.3.4
langchain-community==0.3.14
redis==5.2.1
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
